Removing all idlehook code
After much investigation, determined that there is no way to safely
construct an idle hook class that allows for arbitrary addition
and removal of hooks, given the constraints of FreeRTOS. In reality,
it's not a terrible loss. It's easy enough to just create a generic
work queue class building on threads.