Make threads joinable. Join them on scheduler end.
The threads which are running when the scheduler is being stopped are
since now canceled and joined. The thread which called vTaskEndScheduler
must be treated in a special way because it mustn't cancel itself.
pthread_t of this thread is saved and then by calling
vPortJoinSchedulerEndCaller from main thread cancels the thread and
joins it.